+++ lib/rest_countries_data.dart
@@ -1,10 +1,10 @@
 export 'src/domain/country_model.dart';
 export 'src/domain/enums/country_fields.dart';
 
-import 'package:rest_countries/src/data/countries_api.dart';
-import 'package:rest_countries/src/domain/country_model.dart';
-import 'package:rest_countries/src/domain/enums/country_fields.dart';
-import 'package:rest_countries/src/repository/countries_repository_impl.dart';
+import 'package:rest_countries_data/src/data/countries_api.dart';
+import 'package:rest_countries_data/src/domain/country_model.dart';
+import 'package:rest_countries_data/src/domain/enums/country_fields.dart';
+import 'package:rest_countries_data/src/repository/countries_repository_impl.dart';
 
 class RestCountries {
   static final CountriesRepositoryImpl _repo =